move subtensor rewrites from basic_opt to subtensor_opt
